# Improve performance and reduce binary size