--- a/.cargo/config
+++ b/.cargo/config
@@ -3,6 +3,7 @@
 rustflags = [
     "--cfg", "procmacro2_semver_exempt",
     "-Z", "thinlto=no",
+    "-C", "target-feature=+sse2",
 ]
  
 rustdocflags = [