--- a/bundler/src/bundler/chunk/circular.rs
+++ b/bundler/src/bundler/chunk/circular.rs
@@ -4,8 +4,7 @@ use hygiene::top_level_ident_folder;
 use std::iter::once;
 use swc_common::DUMMY_SP;
 use swc_ecma_ast::*;
-use swc_ecma_transforms::noop_visit_type;
-use swc_ecma_visit::{FoldWith, Node, Visit, VisitWith};
+use swc_ecma_visit::{FoldWith, Node, Visit, VisitWith, noop_visit_type};
 
 mod hygiene;