--- a/bundler/src/bundler/chunk/circular.rs
+++ b/bundler/src/bundler/chunk/circular.rs
@@ -192,6 +192,7 @@ fn merge_respecting_order(mut entry: Vec<ModuleItem>, mut dep: Vec<ModuleItem>) 
     new
 }
 
+/// Searches for top level declaration which provides requirements for `deps`.
 fn dependency_index<T>(item: &ModuleItem, deps: &[T]) -> Option<usize>
 where
     T: Borrow<ModuleItem>,
@@ -248,7 +249,20 @@ where
         }
     }
 
+    fn visit_expr(&mut self, e: &Expr, _: &dyn Node) {
+        match e {
+            // Identifier reference is not a top-level declaration
+            Expr::Ident(..) => return,
+            _ => e.visit_children_with(self),
+        }
+    }
+
     fn visit_class_member(&mut self, _: &ClassMember, _: &dyn Node) {}
     fn visit_function(&mut self, _: &Function, _: &dyn Node) {}
     fn visit_arrow_expr(&mut self, _: &ArrowExpr, _: &dyn Node) {}
+
+    /// We only search for top-level binding
+    fn visit_stmts(&mut self, _: &[Stmt], _: &dyn Node) {}
+    /// We only search for top-level binding
+    fn visit_block_stmt(&mut self, _: &BlockStmt, _: &dyn Node) {}
 }