--- a/bundler/src/bundler/chunk/circular/hygiene.rs
+++ b/bundler/src/bundler/chunk/circular/hygiene.rs
@@ -2,6 +2,7 @@ use crate::id::Id;
 use std::collections::HashSet;
 use swc_common::{Mark, SyntaxContext};
 use swc_ecma_ast::*;
+use swc_ecma_transforms::noop_fold_type;
 use swc_ecma_visit::{Fold, FoldWith};
 
 pub fn top_level_ident_folder(top_level_mark: Mark, module_mark: Mark) -> impl 'static + Fold {
@@ -87,6 +88,8 @@ impl<'a> MergeFolder<'a> {
 }
 
 impl Fold for MergeFolder<'_> {
+    noop_fold_type!();
+
     fn fold_class_decl(&mut self, c: ClassDecl) -> ClassDecl {
         ClassDecl {
             ident: self.fold_bindine_ident(c.ident),