--- a/bundler/src/bundler/chunk/export.rs
+++ b/bundler/src/bundler/chunk/export.rs
@@ -6,8 +6,7 @@ use swc_atoms::js_word;
 use swc_common::{Spanned, SyntaxContext, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_utils::find_ids;
-use swc_ecma_visit::{FoldWith, VisitMut, VisitMutWith};
-use swc_ecma_transforms::noop_visit_mut_type;
+use swc_ecma_visit::{FoldWith, VisitMut, VisitMutWith, noop_visit_mut_type};
 
 impl<L, R> Bundler<'_, L, R>
 where