--- a/ecmascript/transforms/src/compat/es2018/optional_catch_binding.rs
+++ b/ecmascript/transforms/src/compat/es2018/optional_catch_binding.rs
@@ -1,5 +1,5 @@
 use swc_ecma_ast::*;
-use swc_ecma_visit::{Fold, FoldWith};
+use swc_ecma_visit::{noop_fold_type, Fold, FoldWith};
 
 struct OptionalCatchBinding;