spack:
 - Optimize resolver

swc_bunder:
 - Parallize merging of reexports
 - Remove useless `clone`s

swc_ecma_utils:
 - Migrate DropSpan to VisitMut

swc_ecma_transforms:
 - Migrate simple transforms to VisitMut
 - Ignore types to reduce binary size
 - Hide actual types so that we can optimize it in future without breaking change

swc_visit:
 - Apply transforms for vector in-place
 - Apply transforms for box in-place


# Context 

I noticed that there's no visit in the flamegraph, and I benchmarked `VisitMut` and `Fold`.
Result shows that `VisitMut` is <del>1.8</del> 10 times faster than `Fold`. (I forgot subtracting time used to clone module)

I think this difference comes from dereferencing (memory move) of box and recreating boxes and vectors.

As swc is full of `Fold`, converting it to `VisitMut` will improve performance drastically.

## benchmark: Recreation of Vec<T>

When I switched `vec.into_iter().map(|v|fold(v)).collect()` to `vec.move_map(|v|fold(v))`

```
test fold_span      ... bench:  15,712,088 ns/iter (+/- 3,411,948)
test visit_mut_span ... bench:  11,202,239 ns/iter (+/- 1,182,448)
```

## benchmark: Memory move and recreation of Box<T>


```
test base_clone           ... bench:  10,156,408 ns/iter (+/- 1,056,857)
test fold_span            ... bench:  14,239,049 ns/iter (+/- 2,798,167)
test fold_span_pnic       ... bench:  15,233,738 ns/iter (+/- 1,885,537)
test visit_mut_span       ... bench:  10,557,954 ns/iter (+/- 1,195,277)
test visit_mut_span_panic ... bench:  10,964,669 ns/iter (+/- 1,115,552)

```



## three.js benchmark

 - Initial: 29MB / 24 sec

(Many works done)

 - After adding fast path to LocalMarker: 22MB / 5.4 sec
- After mapping vector in place: 20MB / 5.2 sec
- After optimizing resolver (spack): 20MB / 4.7 sec
- After removing dep.clone() while merging: 20MB / 4.5 sec
- After removing dep.clone() while merging reexports: 4.3 sec

(I moved to my linux desktop to profile cpu usage)

Starting point for linux: 20.64MB / 5.49 sec
- After using HashSet<Id> instead of Vec<Id> in LocalMarket: 20.63MB / 4.5 sec
- After optimizing `dependency_index`: 20.63MB / 4.25 sec
 - After migrating dce::UsedMarkRemover to VisitMut: 20.45MB / 4.12 sec
 - After migrating to box.map: 20.44MB / 3.84 sec


From mac:

 - After applying `noop_fold_type` to more passes: 18.66MB / 3.84 sec